Move to py2js file
